create vault page
